site.mk: add ebtables-source-filter
This should prevent some errors in combination with the use of multidomian.

We disabled this to get more flexible, but with our migration to
multidomain it could easily (and in fact already did happen) that
the Prefixes of the wrong domain get announced meshwide. I don't see
another option then to reenable ebtables-source-filter.