ci: build: add target job for a required build status check
It doesn't seem possible to define a matrix as a target for required status checks.
So we need to work arround this by a job which is run after our build matrix.

This accomplishes that.

Required Status Checks help us to ensures that PRs with a failed CI action aren't accidentaly meged.

This also enables us to enable the auto merge feature.

Auto Merge should only be enabled on PRs from trusted sources since it isn't automatically disabled if the source branch has been changed.

For example with our "Bump Gluon" it should be fine to use.